Deleted extra class
