+++ lib/src/core/ui/res/typography/typography.dart
@@ -1,3 +1,81 @@
+import 'package:flutter/material.dart';
+
 abstract class AppTypography {
   
+  static const body1 =  TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w400,
+    fontSize: 16.0,
+    height: 20.0,
+  );
+
+  static const body2 = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w500,
+    fontSize: 18.0,
+    height: 20.0,
+  );
+
+  static const subtitle1 = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w500,
+    fontSize: 16.0,
+    height: 20.0,
+  );
+
+  static const label = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w400,
+    fontSize: 12.0,
+    height: 20.0,
+  );
+
+  static const headline2 = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w600,
+    fontSize: 24.0,
+    height: 32.0,
+  );
+
+  static const caption = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w400,
+    fontSize: 16.0,
+    height: 20.0,
+  );
+
+  static const newCaption = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w500,
+    fontSize: 12.0,
+    height: 14.0,
+  );
+
+  static const button = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w600,
+    fontSize: 16.0,
+    height: 20.0,
+  );
+
+  static const newLargeTitle = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w700,
+    fontSize: 24.0,
+    height: 36.0,
+  );
+
+  static const newSmallTitle = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w600,
+    fontSize: 18.0,
+    height: 20.0,
+  );
+
+  static const promoBold = TextStyle(
+    fontFamily: 'Montserrat',
+    fontWeight: FontWeight.w700,
+    fontSize: 16.0,
+    height: 20.0,
+  );
 }
